update: stop propagation from interactive text to parents.
